requirements.txt
azure_storage==0.37.0
boto3==1.24.26
fastai==2.7.7
matplotlib==3.3.4
msrest==0.7.1
numpy==1.20.1
opencv_python==4.4.0.46
pandas==1.2.4
pdf2image==1.16.0
Pillow==9.2.0
protobuf==4.21.5
pyspellchecker==0.6.3
reportlab==3.6.11
requests==2.25.1
scikit_image==0.18.1
skimage==0.0
spellchecker==0.4
tensorflow==2.4.0
torch==1.12.0
